Converted seel and dewgong to x3d/blender for animation
The seel and dewgong used features which tabula is un-able to animate
properly, so I had to convert to obj, import to blender, assign
childing, and export as x3d.  This is not the fault of the modeller, it
is the fault of tabula not supporting scaling on child parts.